added

+++ msip-bhu-business-impl/msip_bhu_business_ds/src/main/java/com/bhu/vas/business/ds/task/facade/TaskFacadeService.java
@@ -58,7 +58,7 @@ public class TaskFacadeService {
 		if(downtask == null) {
 			throw new BusinessI18nCodeException(ResponseErrorCode.TASK_UNDEFINED,new String[]{String.valueOf(taskid)});
 		}
-		/*if(WifiDeviceDownTask.State_Done.equals(state) || WifiDeviceDownTask.State_Failed.equals(state)){
+		if(WifiDeviceDownTask.State_Done.equals(state) || WifiDeviceDownTask.State_Failed.equals(state) || WifiDeviceDownTask.State_Error.equals(state)){
 			WifiDeviceDownTaskCompleted completed = WifiDeviceDownTaskCompleted.fromWifiDeviceDownTask(downtask, state, response);
 			WifiDeviceDownTaskCompleted result = wifiDeviceDownTaskCompletedService.insert(completed);
 			wifiDeviceDownTaskService.deleteById(taskid);
@@ -67,11 +67,11 @@ public class TaskFacadeService {
 			downtask.setState(state);
 			wifiDeviceDownTaskService.update(downtask);
 			return null;
-		}*/
-		WifiDeviceDownTaskCompleted completed = WifiDeviceDownTaskCompleted.fromWifiDeviceDownTask(downtask, state, response);
+		}
+		/*WifiDeviceDownTaskCompleted completed = WifiDeviceDownTaskCompleted.fromWifiDeviceDownTask(downtask, state, response);
 		WifiDeviceDownTaskCompleted result = wifiDeviceDownTaskCompletedService.insert(completed);
 		wifiDeviceDownTaskService.deleteById(taskid);
-		return result;
+		return result;*/
 	}
 	
 	public void taskComming(WifiDeviceDownTask downtask){
